# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompanyCreate": {
        "description": "会社の新規作成用スキーマ。",
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "department": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "fax": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fax"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "postalCode": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postalcode"
          }
        },
        "required": [
          "name"
        ],
        "title": "CompanyCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "現場報告管理システム Citadel のバックエンド API",
    "title": "Citadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/companies": {
      "get": {
        "operationId": "list_items_api_companies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Items",
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "create_item_api_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "companies"
        ]
      }
    },
...
    "/api/instruments": {
      "get": {
        "operationId": "list_items_api_instruments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Items",
        "tags": [
          "instruments"
        ]
      },
      "post": {
        "operationId": "create_item_api_instruments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstrumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "instruments"
        ]
      }
    },
...
    "/api/sync/delta": {
      "get": {
        "description": "差分同期: 指定日時以降に更新されたレポートとその関連データを取得。\n\n- since: この日時以降に更新されたデータを取得（ISO 8601形式）\n- include_master: Trueの場合、マスタデータも全件含める\n\nレスポンスには `_meta` フィールドが含まれ、同期情報を提供します。",
        "operationId": "delta_sync_api_sync_delta_get",
        "parameters": [
          {
            "description": "ISO 8601形式の日時（例: 2024-01-01T00:00:00Z）",
            "in": "query",
            "name": "since",
            "required": true,
            "schema": {
              "description": "ISO 8601形式の日時（例: 2024-01-01T00:00:00Z）",
              "title": "Since",
              "type": "string"
            }
          },
          {
            "description": "マスタデータも含めるか",
            "in": "query",
            "name": "include_master",
            "required": false,
            "schema": {
              "default": false,
              "description": "マスタデータも含めるか",
              "title": "Include Master",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delta Sync",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/delta/stream": {
      "get": {
        "description": "差分同期を NDJSON でストリーミング配信する。\n各行は {\"table\": \"テーブル名\", \"rows\": [...]}。最後に type: \"meta\" の行で reportCount 等を送る。",
        "operationId": "delta_sync_stream_api_sync_delta_stream_get",
        "parameters": [
          {
            "description": "ISO 8601形式の日時（例: 2024-01-01T00:00:00Z）",
            "in": "query",
            "name": "since",
            "required": true,
            "schema": {
              "description": "ISO 8601形式の日時（例: 2024-01-01T00:00:00Z）",
              "title": "Since",
              "type": "string"
            }
          },
          {
            "description": "マスタデータも含めるか",
            "in": "query",
            "name": "include_master",
            "required": false,
            "schema": {
              "default": false,
              "description": "マスタデータも含めるか",
              "title": "Include Master",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delta Sync Stream",
        "tags": [
          "sync"
        ]
      }
    },
...
    "/api/sync/export": {
      "post": {
        "description": "カスタム条件でデータをエクスポートし、JSON ファイルとしてダウンロードさせる。\n- マスタ: 選択されたテーブルのみ出力\n- レポート: 指定IDのレポートのみ出力\n- モード: edit（ID維持）or copy（IDクリア）",
        "operationId": "export_custom_sync_data_api_sync_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Custom Sync Data",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/handoff": {
      "post": {
        "description": "Direct Handoff: エクスポートとステージングを一度に行う。\nAdmin → Scout への直接転送用。任務（Mission + MissionReport）を発行し、_mission メタデータを付与する。",
        "operationId": "direct_handoff_api_sync_handoff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Direct Handoff",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/stage": {
      "post": {
        "description": "Scout への直接転送用にデータを一時保存する。\nユニークな ticket_id を発行して返す。",
        "operationId": "stage_data_api_sync_stage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage Data",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/stage/{ticket_id}": {
      "get": {
        "description": "ステージングされたデータを取得する。\n取得後、ファイルは削除される（ワンタイム）。",
        "operationId": "get_staged_data_api_sync_stage__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Staged Data",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/upload": {
      "post": {
        "description": "db.json 形式の JSON を受け取り、PostgreSQL に保存する。\nbody に _mission が含まれる場合、任務状態を確認し Purged/Expired なら 403 を返す。\nmode=copy（デフォルト）: レポートは新規UUIDで保存。mode=overwrite: 入力のidが既存なら削除後に同一IDで再登録。\n1 回のリクエストは 1 トランザクション。失敗時はロールバック。",
        "operationId": "upload_data_api_sync_upload_post",
        "parameters": [
          {
            "description": "copy: 新規保存（ID再採番）。overwrite: 同一IDのレポートを上書き。",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "copy",
              "description": "copy: 新規保存（ID再採番）。overwrite: 同一IDのレポートを上書き。",
              "enum": [
                "copy",
                "overwrite"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Data",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/upload/begin": {
      "post": {
        "description": "アップロードセッションを開始する。sessionId と expectedOrder を返す。\n_mission が渡され且つ Purged/Expired の場合は 403 を返す。",
        "operationId": "upload_begin_api_sync_upload_begin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBeginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Begin",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/upload/chunk": {
      "post": {
        "description": "セッションにチャンクを保存する。sequenceIndex は 0 始まりで順番に送ること。\n同一 sequenceIndex の再送は冪等（上書き保存）。",
        "operationId": "upload_chunk_api_sync_upload_chunk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadChunkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Chunk",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/sync/upload/commit": {
      "post": {
        "description": "全チャンクを結合して run_import を実行し、成功時にセッションを削除する。",
        "operationId": "upload_commit_api_sync_upload_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Commit",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}